# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Mock Interview Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /resumes:
    get:
      operationId: listResumes
      tags: [resumes]
      summary: List user's resumes
      responses:
        "200":
          description: Resume list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resume"
    post:
      operationId: uploadResume
      tags: [resumes]
      summary: Upload a resume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeInput"
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
  /resumes/{id}:
    get:
      operationId: getResume
      tags: [resumes]
      summary: Get a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
    delete:
      operationId: deleteResume
      tags: [resumes]
      summary: Delete a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Interviews
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        avatarUrl:
          type: ["string", "null"]
        targetRole:
          type: ["string", "null"]
        experienceLevel:
          type: ["string", "null"]
        createdAt:
          type: string
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: string
        targetRole:
          type: string
        experienceLevel:
          type: string
...
    Resume:
      type: object
      required: [id, userId, filename, uploadedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        filename:
          type: string
        content:
          type: ["string", "null"]
        parsedSkills:
          type: ["string", "null"]
        isActive:
          type: boolean
        uploadedAt:
          type: string
    ResumeInput:
      type: object
      required: [filename, content]
      properties:
        filename:
          type: string
        content:
          type: string

    # Interviews
...